requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Background Tasks
celery==5.3.4
flower==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# File Processing
python-magic==0.4.27
Pillow==10.1.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2

# AI/ML & NLP
openai==1.3.7
anthropic==0.7.8
transformers==4.36.0
torch==2.1.1
numpy==1.24.3
pandas==2.1.4

# Search & Indexing
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0

# Cloud Services
boto3==1.34.0
google-cloud-storage==2.10.0
google-cloud-vision==3.4.6

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development & Monitoring
python-dotenv==1.0.0
structlog==23.2.0
prometheus-client==0.19.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
